% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{draw_confirmed}
\alias{draw_confirmed}
\title{Create a two panel plot of COVID-19, by count and by density of confirmed
cases.}
\usage{
draw_confirmed(x = merge_census(), N = 1000)
}
\arguments{
\item{x}{sf object of merged COVID-19 and census data}

\item{N}{numeric, scaling for density.  By default cases per 1000 people.}
}
\value{
a list of 2 ggplot2 graphics.  Use patchwork to draw together ala
    \code{result[[1]] + result[[2]]}
}
\description{
Create a two panel plot of COVID-19, by count and by density of confirmed
cases.
}
