% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{draw_history}
\alias{draw_history}
\title{Create a plot of COVID-19, by count and by density of confirmed
cases.}
\usage{
draw_history(
  x = read_datahub(state = "Maine"),
  pop = read_census(state = "Maine"),
  varname = "Confirmed",
  State_name = "",
  N = 1,
  threshold = c(10, NA)[1],
  logscale = TRUE
)
}
\arguments{
\item{x}{tibble of COVID-19 and census data}

\item{pop}{SF MUTLIPOLYGON table of county level geometry and population}

\item{varname}{character, the name of the variable to draw ala
'Confirmed', 'Recovered', 'Deaths' etc.}

\item{State_name}{character, for title, otherwise skip with ""}

\item{N}{numeric, scaling for density.  By default cases per 1000 people.}

\item{threshold}{numeric, clip records (by county) where varname is less than this
value.  Set to NA to keep all.}

\item{logscale}{logical, if TRUE then logscale the the dependent variables}
}
\value{
a ggplot object
}
\description{
Create a plot of COVID-19, by count and by density of confirmed
cases.
}
