% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{draw_infected}
\alias{draw_infected}
\title{Create a two panel plot of COVID-19, by count and by density of confirmed
cases.}
\usage{
draw_infected(
  x = read_datahub(state = "Maine"),
  pop = read_census(state = "Maine"),
  N = 1000,
  logscale = TRUE
)
}
\arguments{
\item{x}{tibble of COVID-19 and census data}

\item{pop}{SF MUTLIPOLYGON table of county level geometry and population}

\item{N}{numeric, scaling for density.  By default cases per 1000 people.}

\item{logscale}{logical, if TRUE then logscale the the dependent variables}
}
\value{
a list of 2 ggplot2 graphics.  Use patchwork to draw together ala
    \code{result[[1]] + result[[2]]}
}
\description{
Create a two panel plot of COVID-19, by count and by density of confirmed
cases.
}
