% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census.R
\name{fetch_census_estimates}
\alias{fetch_census_estimates}
\title{Get census estimates for counties by state}
\usage{
fetch_census_estimates(
  state = "all",
  simplify = TRUE,
  save_file = TRUE,
  path = stcovid_path("census")
)
}
\arguments{
\item{state}{character, the name of the state or 'all'}

\item{simplify}{logical, if TRUE simplify into a familiar form}

\item{save_file}{logical if TRUE save as geojson, ignored if simplify = FALSE}

\item{path}{character, the path to save the file}
}
\value{
sf table of county population, density, geoid and geometry.
        If \code{state} is 'all' then iterate through all states and return
        a list of sf tables
}
\description{
Get census estimates for counties by state
}
