% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid19.R
\name{fetch_datahub}
\alias{fetch_datahub}
\title{Fetch updated county data from COVID19 data hub}
\usage{
fetch_datahub(
  simplify = TRUE,
  save_file = TRUE,
  path = stcovid_path("cumcounts"),
  ...
)
}
\arguments{
\item{simplify}{logical, if TRUE trim the dataset}

\item{save_file}{logical, if TRUE save the file.  Ignored if simplify is FALSE.}

\item{path}{charcater, the path to save the data}

\item{...}{further arguments for \code{\link[COVID19]{covid19}}}
}
\value{
tibble of COVID19 data hub data
}
\description{
Fetch updated county data from COVID19 data hub
}
\seealso{
\url{https://covid19datahub.io/}
}
