% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census.R
\name{read_census}
\alias{read_census}
\title{Read locally stored state population, density and boundaries from US Census}
\usage{
read_census(state = "Maine", path = stcovid_path("census"))
}
\arguments{
\item{state}{character, the name of the state}

\item{path}{character, the path to the data file}
}
\value{
sf MULTIPOLYGIN object of county populations and boundaries
}
\description{
Read locally stored state population, density and boundaries from US Census
}
