% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid19.R
\name{read_datahub}
\alias{read_datahub}
\title{Read the locally store datahub table}
\usage{
read_datahub(
  filename = stcovid_path("cumcounts", "datahub.csv"),
  state = NULL,
  date = NULL
)
}
\arguments{
\item{filename}{character, the name of the data file}

\item{state}{NULL or character, to filter the data}

\item{date}{NULL or Date, to filter the data}
}
\value{
tibble
}
\description{
Read the locally store datahub table
}
