% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{st_abb2name}
\alias{st_abb2name}
\title{Convert state abbreviation to full name}
\usage{
st_abb2name(x = state.abb)
}
\arguments{
\item{x}{character, one or more state abbreviations}
}
\value{
full state names or NA when not found
}
\description{
Convert state abbreviation to full name
}
