% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{st_name2abb}
\alias{st_name2abb}
\title{Convert state names to abbreviations}
\usage{
st_name2abb(x = state.name)
}
\arguments{
\item{x}{character, one or more state names}
}
\value{
state abbreviations or NA when not found
}
\description{
Convert state names to abbreviations
}
