% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stcovid-package.R
\docType{package}
\name{stcovid}
\alias{stcovid}
\title{stcovid package}
\description{
Read and filter [COVID-19](https://github.com/covid19datahub/COVID19) data offerings,
with a county-by-sate centric perspective.
}
