% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid19.R
\name{update_datahub}
\alias{update_datahub}
\title{Update the locally stored datahub table}
\usage{
update_datahub(filename = stcovid_path("cumcounts", "datahub.csv"))
}
\arguments{
\item{filename}{character, the name of the data file with datahub data}
}
\value{
tibble of COVID19 data hub data
}
\description{
Update the locally stored datahub table
}
